refactor(analysis): extract aggregation logic to snapshot-aggregator module
domain-analysis-snapshot-builder.ts was 484 lines (21% over the 400-line
limit). Extracted Phase 1/2 accumulation + normalisation logic and all
low-level helper functions into a new domain-analysis-snapshot-aggregator.ts
(270 lines). Builder is now 250 lines and imports aggregateAnalysisRows().
No behaviour change — pure structural split.
